add signal generator, add SNR test, add truncated w(4x2) performance test.

+++ fixpoint_matlab_simulator/w_bypassed.m
@@ -3,7 +3,8 @@ function [w] = w_bypassed(w,bypass)
 
 group_num=2^nnz(bypass);
 w=w(1:group_num:end,:);
-w=int16(kron(double(w),ones(group_num,1)));
+% w=int16(kron(double(w),ones(group_num,1)));
+w=kron(double(w),ones(group_num,1));
 
 end
